+++ src/bind.cpp
@@ -135,15 +135,17 @@ List cbind__impl( Dots dots ){
 
   std::vector<DataFrameAble> chunks ;
   for( int i=0; i<n; i++) {
+    if( ! Rf_isNull(dots[i]) )
     chunks.push_back( DataFrameAble( dots[i] ) );
   }
+  n = chunks.size() ;
 
   // first check that the number of rows is the same
   const DataFrameAble& df = chunks[0] ;
   int nrows = df.nrows() ;
   int nv = df.size() ;
   for( int i=1; i<n; i++){
-    const DataFrameAble& current = dots[i] ;
+    const DataFrameAble& current = chunks[i] ;
     if( current.nrows() != nrows ){
       stop( "incompatible number of rows (%d, expecting %d)", current.nrows(), nrows ) ;
     }
@@ -158,7 +160,7 @@ List cbind__impl( Dots dots ){
   for( int i=0, k=0 ; i<n; i++){
       Rcpp::checkUserInterrupt() ;
 
-      const DataFrameAble& current = dots[i] ;
+      const DataFrameAble& current = chunks[i] ;
       CharacterVector current_names = current.names() ;
       int nc = current.size() ;
       for( int j=0; j<nc; j++, k++){